Move stability decisions to their own class
