+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj
@@ -214,6 +214,7 @@
     <ClCompile Include="Source\HOI4World\Decisions\Decisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\DecisionsCategory.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\IdeologicalDecisions.cpp" />
+    <ClCompile Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Division.cpp" />
     <ClCompile Include="Source\HOI4World\DivisionTemplate.cpp" />
     <ClCompile Include="Source\HOI4World\Event.cpp" />
@@ -349,6 +350,7 @@
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsCategories.h" />
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsCategory.h" />
     <ClInclude Include="Source\HOI4World\Decisions\IdeologicalDecisions.h" />
+    <ClInclude Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.h" />
     <ClInclude Include="Source\HOI4World\Division.h" />
     <ClInclude Include="Source\HOI4World\DivisionTemplate.h" />
     <ClInclude Include="Source\HOI4World\Event.h" />
